requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
credentials==1.1
cryptography==45.0.4
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
executing==2.2.0
fabric==3.2.2
frozenlist==1.7.0
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gspread==6.2.1
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.10
invoke==2.2.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
multidict==6.4.4
nest-asyncio==1.6.0
oauthlib==3.2.2
openai==0.28.0
packaging==24.2
paramiko==3.5.1
parso==0.8.4
platformdirs==4.3.7
prompt_toolkit==3.0.50
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyNaCl==1.5.0
pyparsing==3.2.3
pyTelegramBotAPI==4.27.0
python-dateutil==2.9.0.post0
python-telegram-bot==20.3
pywin32==310
pyzmq==26.4.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
style==1.1.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
update==0.0.1
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.20.1
